Replace entity liveloading templates
