+++ modules/core/entity/entity_embeds.js
@@ -1,0 +1,62 @@
+/**
+ * Implements hook_frontend_embed
+ * Process entity embeds
+ */
+
+iris.modules.entity.registerHook("hook_frontend_embed__entity", 0, function (thisHook, data) {
+
+  iris.invokeHook("hook_entity_fetch", thisHook.authPass, thisHook.context.embedOptions, thisHook.context.embedOptions).then(function (result) {
+
+      if (thisHook.context.embedOptions.variableName) {
+
+        thisHook.context.embedID = thisHook.context.embedOptions.variableName;
+
+      } else {
+
+        thisHook.context.embedID = JSON.stringify(thisHook.context.embedOptions);
+
+      }
+
+      thisHook.context.vars.tags.headTags["socket.io"] = {
+        type: "script",
+        attributes: {
+          "src": "/socket.io/socket.io.js"
+        },
+        rank: -1
+      }
+
+      thisHook.context.vars.tags.headTags["handlebars"] = {
+        type: "script",
+        attributes: {
+          "src": "/modules/entity/handlebars.min.js"
+        },
+        rank: -1
+      }
+
+      thisHook.context.vars.tags.headTags["entity_fetch"] = {
+        type: "script",
+        attributes: {
+          "src": "/modules/entity/templates.js"
+        },
+        rank: 0
+      };
+
+      var entityPackage = "\n" + "iris.entityPreFetch(" + JSON.stringify(result) + ", '" + thisHook.context.embedID + "'" + ", " + JSON.stringify(thisHook.context.embedOptions) + ")";
+
+      var loader = entityPackage;
+
+      thisHook.pass({
+        html: "<script>" + loader + "</script>",
+        variables: [result],
+        blockHeader: "<script>" + loader + "</script>",
+        blockFooter: ""
+      });
+
+    },
+    function (error) {
+
+      thisHook.pass(data);
+
+    });
+
+});
