+++ modules/core/entity/entity_views.js
@@ -4,68 +4,21 @@
 
 var fs = require("fs");
 
-/**
- * Implements hook_frontend_embed
- * Process entity embeds
- */
+require("./entity_embeds");
 
-iris.modules.entity.registerHook("hook_frontend_embed__entity", 0, function (thisHook, data) {
+var updateEmbeds = function () {
 
-  iris.invokeHook("hook_entity_fetch", thisHook.authPass, thisHook.context.embedOptions, thisHook.context.embedOptions).then(function (result) {
+  if (iris.modules.frontend.globals.liveEmbeds.entity) {
 
-      if (thisHook.context.embedOptions.variableName) {
+    Object.keys(iris.modules.frontend.globals.liveEmbeds.entity).forEach(function (embed) {
 
-        thisHook.context.embedID = thisHook.context.embedOptions.variableName;
+      iris.modules.frontend.globals.liveEmbeds.entity[embed].sendResult();
 
-      } else {
+    })
 
-        thisHook.context.embedID = JSON.stringify(thisHook.context.embedOptions);
+  }
 
-      }
-
-      thisHook.context.vars.tags.headTags["socket.io"] = {
-        type: "script",
-        attributes: {
-          "src": "/socket.io/socket.io.js"
-        },
-        rank: -1
-      }
-
-      thisHook.context.vars.tags.headTags["handlebars"] = {
-        type: "script",
-        attributes: {
-          "src": "/modules/entity/handlebars.min.js"
-        },
-        rank: -1
-      }
-
-      thisHook.context.vars.tags.headTags["entity_fetch"] = {
-        type: "script",
-        attributes: {
-          "src": "/modules/entity/templates.js"
-        },
-        rank: 0
-      };
-
-      var entityPackage = "\n" + "iris.entityPreFetch(" + JSON.stringify(result) + ", '" + thisHook.context.embedID + "'" + ", " + JSON.stringify(thisHook.context.embedOptions) + ")";
-
-      var loader = entityPackage;
-
-      thisHook.pass({
-        html: "<script>" + loader + "</script>",
-        variables: [result],
-        blockHeader: "<script>" + loader + "</script>",
-        blockFooter: ""
-      });
-
-    },
-    function (error) {
-
-      thisHook.pass(data);
-
-    });
-
-});
+};
 
 /**
  * @member hook_entity_created
@@ -76,6 +29,8 @@ iris.modules.entity.registerHook("hook_frontend_embed__entity", 0, function (thi
  * This hook is run once an entity has been created; useful for live updates or keeping track of changes
  */
 iris.modules.entity.registerHook("hook_entity_created", 0, function (thisHook, entity) {
+  
+  updateEmbeds();
 
   for (var authUser in iris.modules.auth.globals.userList) {
 
@@ -116,6 +71,8 @@ iris.modules.entity.registerHook("hook_entity_created", 0, function (thisHook, e
  * This hook is run when an entity is updated/edited; useful for live updates or keeping track of changes
  */
 iris.modules.entity.registerHook("hook_entity_updated", 0, function (thisHook, entity) {
+  
+  updateEmbeds();
 
   for (var authUser in iris.modules.auth.globals.userList) {
 
@@ -156,6 +113,8 @@ iris.modules.entity.registerHook("hook_entity_updated", 0, function (thisHook, e
  * This hook is run when an entity is deleted; useful for live updates or keeping track of changes
  */
 iris.modules.entity.registerHook("hook_entity_deleted", 0, function (thisHook, entity) {
+  
+  updateEmbeds();
 
   for (var authUser in iris.modules.auth.globals.userList) {
